added receiving for create-room
